Refactor agent flow and improve tool coverage
