+++ .env.local.example
@@ -1,6 +1,20 @@
-AGENT_API_SECRET=your-secret-here
-# Optional for local browser access to the protected /api/agent route.
-# In local dev, set this to the same value as AGENT_API_SECRET.
-NEXT_PUBLIC_AGENT_API_SECRET=your-secret-here
+# Required: server-side model access.
+DEEPSEEK_API_KEY=your-deepseek-api-key
+
+# Optional: DeepSeek-compatible API settings.
+DEEPSEEK_BASE_URL=https://api.deepseek.com
+DEEPSEEK_MODEL=deepseek-v4-flash
+
+# Required: protects POST /api/agent on the server.
+AGENT_API_SECRET=replace-with-a-long-random-local-secret
+
+# Local dev only: this value is exposed to the browser because of NEXT_PUBLIC.
+# Set it to the same value as AGENT_API_SECRET only for local development.
+NEXT_PUBLIC_AGENT_API_SECRET=replace-with-the-same-local-secret
+
+# Optional: absolute folder the agent may inspect and draft edits inside.
+# Leave empty to use data/workspace.
+# AGENT_WORKSPACE_ROOT=C:\path\to\project
+
 # Optional: set this only if GitHub CLI is not available on PATH.
-GH_PATH=C:\path\to\gh.exe
+# GH_PATH=C:\path\to\gh.exe
