+++ eslint.config.mjs
@@ -1,0 +1,17 @@
+import { FlatCompat } from "@eslint/eslintrc";
+import { dirname } from "node:path";
+import { fileURLToPath } from "node:url";
+
+const __dirname = dirname(fileURLToPath(import.meta.url));
+const compat = new FlatCompat({
+  baseDirectory: __dirname,
+});
+
+const eslintConfig = [
+  ...compat.extends("next/core-web-vitals", "next/typescript"),
+  {
+    ignores: [".next/**", ".next-dev/**", ".test-dist/**", "next-env.d.ts", "node_modules/**"],
+  },
+];
+
+export default eslintConfig;
